fix: Update surprisal calculator to batch inputs

+++ Scripts/utils.py
@@ -64,11 +64,10 @@ def to_tokens_and_logprobs(model, tokenizer, input_texts):
     print("Creating surprisal tables:")
     batch = []
     sentence_id = 0
-    for i, id_surp in enumerate(zip(input_ids, gen_surprisals)):
+    for i, id_surp in tqdm(enumerate(zip(input_ids, gen_surprisals))):
         sentence = []
         input_sentence, input_surprisals = id_surp
-        print(f"Processing document {i}")
-        for token, p in tqdm(zip(input_sentence, input_surprisals)):
+        for token, p in zip(input_sentence, input_surprisals):
             if token not in tokenizer.all_special_ids:
                 sentence.append({
                     # "sentence_id": sentence_id,
@@ -99,18 +98,22 @@ def scrape(article_urls):
     return pot
 
 def prompt_hfds(num_articles,client, temperature, model_name, 
-                model_output_dir="../Generations", human_output_dir="../Sources", write_source=True):
+                model_output_dir="../Generations", human_output_dir="../Sources", 
+                write_source=True, regenerate=True):
     data = load_dataset("abisee/cnn_dailymail", "1.0.0", trust_remote_code=True)['train']
     shuffled_data = data.shuffle()
     if num_articles < 0:
         num_articles = 10_000
     for i in tqdm(range(num_articles)):
+        generation_output_path = Path(model_output_dir) / Path(f"{model_name}_{i}.txt")
+        source_output_path = Path(human_output_dir) / Path(f"human_{i}.txt")
+        if not regenerate and generation_output_path.exists():
+            continue
         article = data[i]
         first_sent = get_first_sentence(article['article'])
         generation = generate(client=client, prompt=first_sent, temperature=temperature, model=model_name)
-        with open(Path(model_output_dir) / Path(f"{model_name}_{i}.txt"), "w") as outfile:
+        with open(generation_output_path, "w") as outfile:
             outfile.write(generation)
-        source_output_path = Path(human_output_dir) / Path(f"human_{i}.txt")
         if not source_output_path.exists() and write_source:
             with open(source_output_path, "w") as outfile:
                 outfile.write(article['article'])
@@ -126,24 +129,39 @@ def generate(client, prompt, temperature, model):
   )
   return completion.choices[0].message.content
 
-def calc_surprisal(model, tokenizer, input_dir, output_dir, num_files=-1):
+def split_batches(lst, batch_size):
+    num_batches = int(np.ceil(len(lst) / batch_size))
+    batched = []
+    for i in range(num_batches):
+        start_idx = i * batch_size
+        batch = lst[start_idx:start_idx + batch_size]
+        batched.append(batch)
+    return batched
+
+def calc_surprisal(model, tokenizer, input_dir, output_dir, num_files=-1, batch_size=20):
     # Setting up files
     root = Path(input_dir)
     text_filepaths = list(root.iterdir())
     absolute_filepaths = [file.resolve() for file in text_filepaths]
 
     # Reading in files
-    texts=[]
+    if num_files > 0:
+        num_files = len(absolute_filepaths + 1)
+    all_texts=[]
     for filepath in tqdm(absolute_filepaths[:num_files]):
         with open(filepath, 'r') as file:
             text = file.read()
-        texts.append(text)
+        all_texts.append(text)
+
+    batched_texts = split_batches(all_texts, batch_size=batch_size)
+    batched_filepaths = split_batches(text_filepaths, batch_size=batch_size)
 
     # calculating surprisals with model
-    batch = to_tokens_and_logprobs(model, tokenizer, texts)
+    for texts, paths in tqdm(zip(batched_texts, batched_filepaths)):
+        output = to_tokens_and_logprobs(model, tokenizer, texts)
 
-    # writing ouput files
-    output_root = Path(output_dir)
-    for fp, text in zip(text_filepaths, batch):
-        output_fp = output_root / fp.with_suffix(".csv").name
-        text.to_csv(output_fp)
+        # writing ouput files
+        output_root = Path(output_dir)
+        for fp, text in zip(paths, output):
+            output_fp = output_root / fp.with_suffix(".csv").name
+            text.to_csv(output_fp)
